+++ views/backend/messageView.php
@@ -2,7 +2,7 @@
 <h1>Messagerie</h1>
 
 <a class="button" href="index.php?action=adminMessages">Retour à la liste des messages</a>
-<a class="button" href="index.php?action=deleteMessage&amp;id=<?= $message['id'] ?>">Supprimer le message</a>
+<a class="button" href="index.php?action=deleteMessage&amp;id=<?= $message['id'] ?>" onclick="if(confirm('Supprimer définitivement ?')){return true;}else{return false;}">Supprimer le message</a>
 
 <section id="message">
 	<h3><?= $message['subject'] ?></h3>
